+++ tidypolars/funs.py
@@ -1,0 +1,43 @@
+import polars as pl
+
+__all__ = ["case_when", "if_else"]
+
+def case_when(expr):
+    """
+    Case when
+
+    Parameters
+    ----------
+    expr : Expr
+        A logical expression
+
+    Examples
+    --------
+    >>> df = tp.Tibble(x = range(1, 4))
+    >>> df.mutate(
+    >>>    case_x = tp.case_when(col('x') < 2).then(1)
+    >>>             .when(col('x') < 3).then(2)
+    >>>             .otherwise(0)
+    >>> )
+    """
+    return pl.when(expr)
+
+def if_else(condition, true, false):
+    """
+    If Else
+
+    Parameters
+    ----------
+    condition : Expr
+        A logical expression
+    true :
+        Value if the condition is true
+    false :
+        Value if the condition is false
+
+    Examples
+    --------
+    >>> df = tp.Tibble(x = range(1, 4))
+    >>> df.mutate(if_x = tp.if_else(col('x') < 2, 1, 2))
+    """
+    return pl.when(condition).then(true).otherwise(false)