+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v0.1.7 (in development)
 * New Tibble methods
+  + `.count()`
   + `.drop_null()`
   + `.inner_join()`/`.left_join()`/`.outer_join()`
   + `.write_csv()`